pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
